Colors, damping, and bugfixes
Heart material color chosen randomly from a select number of choices.

Animation used to stop after tab lost focus - this was fixed, I think, by resetting the position based on a position check and not only opacity. In other words, if the object is way out of bounds, reset its position. For some reason, opacity does not update when the animation loop is stopped, so it is not a reliable check when the tab is out of focus.

The OrbitControls now has some inertia to it, which matches the feel of the scene.